add create page and lg

+++ OmniLine/Models/CounterAgent.cs
@@ -5,7 +5,7 @@ namespace OmniLine.Models
     public class CounterAgent
     {
         [Key]
-        public long CounterAgentId { get; set; }
+        public int CounterAgentId { get; set; }
         public string Name { get; set; }
         public string DateCreate { get; set; }
         public string DateEdit { get; set; }
